+++ lang/test/org/partiql/lang/eval/builtins/functions/TrimEvaluationTest.kt
@@ -3,12 +3,16 @@ package org.partiql.lang.eval.builtins.functions
 import org.junit.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ArgumentsSource
+import org.partiql.lang.errors.ErrorCode
+import org.partiql.lang.errors.Property
 import org.partiql.lang.eval.EvaluatorTestBase
 import org.partiql.lang.eval.builtins.Argument
 import org.partiql.lang.eval.builtins.ExprFunctionTestCase
 import org.partiql.lang.eval.builtins.checkInvalidArgType
 import org.partiql.lang.types.StaticType
 import org.partiql.lang.util.ArgumentsProviderBase
+import org.partiql.lang.util.propertyValueMapOf
+import org.partiql.lang.util.to
 
 class TrimEvaluationTest : EvaluatorTestBase() {
     // Pass test cases
@@ -67,9 +71,26 @@ class TrimEvaluationTest : EvaluatorTestBase() {
             ExprFunctionTestCase("trim(both '' from missing)", "null"),
             ExprFunctionTestCase("trim(missing from '')", "null"),
             ExprFunctionTestCase("trim('' from missing)", "null"),
-            ExprFunctionTestCase("trim(missing)", "null")
+            ExprFunctionTestCase("trim(missing)", "null"),
+            // test for upper case trim spec
+            ExprFunctionTestCase("trim(BOTH from '   string   ')", "\"string\""),
+            ExprFunctionTestCase("trim(LEADING from '   string   ')", "\"string   \""),
+            ExprFunctionTestCase("trim(TRAILING from 'string   ')", "\"string\""),
         )
     }
+
+    // Consider trim(something from ' string '), here "something" will be interpreted as an identifier
+    @Test
+    fun trimSubstringNoBinding() = runEvaluatorErrorTestCase(
+        query = "trim(something from ' string ')",
+        expectedErrorCode = ErrorCode.EVALUATOR_BINDING_DOES_NOT_EXIST,
+        expectedErrorContext = propertyValueMapOf(1, 6, Property.BINDING_NAME to "something"),
+        expectedPermissiveModeResult = "MISSING"
+    )
+
+    // Consider trim(something ' ' from ' string '), here "something" will be interpreted as a trim specification
+    // this test case is tested currently in ParserErrorsTests
+    // todo: decide where should we catch the error and clean up the comments
 
     // Error test cases: Invalid argument type
     @Test